+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Added `uuid` to `Account` [#14](https://github.com/recurly/recurly-client-php-internal/pull/14)
 * Added `redemptions` to `Account` [#14](https://github.com/recurly/recurly-client-php-internal/pull/14)
 * Add `max_redemptions_per_account` to `Coupon`
+* Add `redemptions` to `Subscription`
 
 ## Version 2.4.5 (August 3, 2015)